+++ src/card/sets/y2011/mod.rs
@@ -1,6 +1,7 @@
 //! Card sets first released in 2011.
 
 pub(super) mod innistrad;
+pub(super) mod magic_2012;
 pub(super) mod mirrodin_besieged;
 
 pub(super) use super::{CardRecord, PrintingRecord};